requirements.txt
# python3 -m venv .venv
python-dotenv
openai
